+++ test/tasks/ods-package-image_test.go
@@ -15,6 +15,7 @@ import (
 	"github.com/opendevstack/pipeline/pkg/logging"
 	"github.com/opendevstack/pipeline/pkg/pipelinectxt"
 	"github.com/opendevstack/pipeline/pkg/tasktesting"
+	"k8s.io/utils/strings/slices"
 )
 
 func TestTaskODSPackageImage(t *testing.T) {
@@ -48,6 +49,26 @@ func TestTaskODSPackageImage(t *testing.T) {
 					wsDir := ctxt.Workspaces["source"]
 					checkResultingFiles(t, ctxt, wsDir)
 					checkResultingImageHelloWorld(t, ctxt, wsDir)
+				},
+			},
+			"task should build image with additional tags": {
+				WorkspaceDirMapping: map[string]string{"source": "hello-world-app"},
+				PreRunFunc: func(t *testing.T, ctxt *tasktesting.TaskRunContext) {
+					wsDir := ctxt.Workspaces["source"]
+					ctxt.ODS = tasktesting.SetupGitRepo(t, ctxt.Namespace, wsDir)
+					ctxt.Params = map[string]string{
+						"extra-tags": "'latest cool'",
+					}
+				},
+				WantRunSuccess: true,
+				PostRunFunc: func(t *testing.T, ctxt *tasktesting.TaskRunContext) {
+					wsDir := ctxt.Workspaces["source"]
+					checkResultingFiles(t, ctxt, wsDir)
+					checkTagFiles(t, ctxt, wsDir, []string{"latest", "cool"})
+					checkTags(t, ctxt, wsDir, []string{ctxt.ODS.GitCommitSHA, "latest", "cool"})
+					checkResultingImageHelloWorld(t, ctxt, wsDir)
+					checkTaggedImageHelloWorld(t, ctxt, wsDir, "latest")
+					checkTaggedImageHelloWorld(t, ctxt, wsDir, "cool")
 				},
 			},
 			"task should reuse existing image": {
@@ -94,10 +115,61 @@ func checkResultingFiles(t *testing.T, ctxt *tasktesting.TaskRunContext, wsDir s
 	}
 }
 
-func runResultingImage(t *testing.T, ctxt *tasktesting.TaskRunContext, wsDir string) string {
+func checkTagFiles(t *testing.T, ctxt *tasktesting.TaskRunContext, wsDir string, tags []string) {
+	wantFiles := []string{}
+	for _, tag := range tags {
+		wantFiles = append(wantFiles, fmt.Sprintf(".ods/artifacts/image-digests/%s-%s.json", ctxt.ODS.Component, tag))
+	}
+	for _, wf := range wantFiles {
+		if _, err := os.Stat(filepath.Join(wsDir, wf)); os.IsNotExist(err) {
+			t.Fatalf("Want %s, but got nothing", wf)
+		}
+	}
+}
+
+func checkTags(t *testing.T, ctxt *tasktesting.TaskRunContext, wsDir string, expectedTags []string) {
+	// registry := "kind-registry.kind:5000"
+	registry := "localhost:5000"
+	tlsVerify := false
+	args := []string{
+		"inspect",
+		`--format={{.RepoTags}}`,
+		fmt.Sprintf("--tls-verify=%v", tlsVerify),
+	}
+	imageNsStreamSha := fmt.Sprintf("%s/%s:%s", ctxt.Namespace, ctxt.ODS.Component, ctxt.ODS.GitCommitSHA)
+	imageRef := fmt.Sprintf("docker://%s/%s", registry, imageNsStreamSha)
+	args = append(args, imageRef)
+
+	stdout, _, err := command.RunBuffered("skopeo", args)
+	if err != nil {
+		t.Fatalf("skopeo inspect %s: %s", fmt.Sprint(args), err)
+	}
+	tags, err := parseSkopeoInspectDigestTags(string(stdout))
+	if err != nil {
+		t.Fatalf("parse tags failed: %s", err)
+	}
+	for _, expectedTag := range expectedTags {
+		if !slices.Contains(tags, expectedTag) {
+			t.Fatalf("Expected tags=%s to be in actual tags=%s", fmt.Sprint(expectedTags), fmt.Sprint(tags))
+		}
+	}
+}
+
+func parseSkopeoInspectDigestTags(out string) ([]string, error) {
+	t := strings.TrimSpace(out)
+	if !(strings.HasPrefix(t, "[") && strings.HasSuffix(t, "]")) {
+		return nil, fmt.Errorf("skopeo inspect: unexpected tag response expecting tags to be in brackets %s", t)
+	}
+	t = t[1 : len(t)-1]
+	// expecting t to have space separated tags.
+	tags := strings.Split(t, " ")
+	return tags, nil
+}
+
+func runSpecifiedImage(t *testing.T, ctxt *tasktesting.TaskRunContext, wsDir string, image string) string {
 	stdout, stderr, err := command.RunBuffered("docker", []string{
 		"run", "--rm",
-		getDockerImageTag(t, ctxt, wsDir),
+		image,
 	})
 	if err != nil {
 		t.Fatalf("could not run built image: %s, stderr: %s", err, string(stderr))
@@ -106,8 +178,22 @@ func runResultingImage(t *testing.T, ctxt *tasktesting.TaskRunContext, wsDir str
 	return got
 }
 
+func runResultingImage(t *testing.T, ctxt *tasktesting.TaskRunContext, wsDir string) string {
+	got := runSpecifiedImage(t, ctxt, wsDir, getDockerImageTag(t, ctxt, wsDir))
+	return got
+}
+
 func checkResultingImageHelloWorld(t *testing.T, ctxt *tasktesting.TaskRunContext, wsDir string) {
 	got := runResultingImage(t, ctxt, wsDir)
+	want := "Hello World"
+	if got != want {
+		t.Fatalf("Want %s, but got %s", want, got)
+	}
+}
+
+func checkTaggedImageHelloWorld(t *testing.T, ctxt *tasktesting.TaskRunContext, wsDir string, tag string) {
+	image := fmt.Sprintf("localhost:5000/%s/%s:%s", ctxt.Namespace, ctxt.ODS.Component, tag)
+	got := runSpecifiedImage(t, ctxt, wsDir, image)
 	want := "Hello World"
 	if got != want {
 		t.Fatalf("Want %s, but got %s", want, got)